+++ Balancer-Bots/Large-Flash-Loan/src/utils.ts
@@ -1,4 +1,4 @@
-import { ethers, Finding, FindingSeverity, FindingType } from "forta-agent";
+import { ethers } from "forta-agent";
 import BigNumber from "bignumber.js";
 import { Mutex } from "async-mutex";
 import LRU from "lru-cache";
@@ -9,25 +9,6 @@ export interface NetworkData {
 }
 
 export type AgentConfig = Record<number, NetworkData>;
-
-export function createFinding(flashLoanLog: ethers.utils.LogDescription, tvlPercentage: BigNumber): Finding {
-  const { recipient, token, amount } = flashLoanLog.args;
-
-  return Finding.from({
-    name: "Large flash loan",
-    description: "A flash loan that involved a significant percentage of a token's balance was detected",
-    alertId: "BAL-4",
-    protocol: "Balancer",
-    type: FindingType.Info,
-    severity: FindingSeverity.Unknown,
-    metadata: {
-      recipient,
-      token,
-      amount: amount.toString(),
-      tvlPercentage: tvlPercentage.toString(10),
-    },
-  });
-}
 
 export const toBn = (ethersBn: ethers.BigNumber) => new BigNumber(ethersBn.toString());
